вынос валидации в модуль, фикс отправки формы
